chore: transaction classification (IsSigningTransaction)
